_legacy/server/requirements.txt
Flask==3.1.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
python-dotenv==1.0.1
psycopg2-binary==2.9.9
requests==2.32.3
